# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Divine Children Home Ltd API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get a single job listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Not found
...
  /news/{slug}:
    get:
      operationId: getNewsArticle
      tags: [news]
      summary: Get a single news article by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Article details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsArticle"
        "404":
          description: Not found
...
components:
  schemas:
...
    Job:
      type: object
      required: [id, title, department, location, type, description, salary, postedAt, isActive]
      properties:
        id:
          type: integer
        title:
          type: string
        department:
          type: string
        location:
          type: string
        type:
          type: string
          description: "full-time, part-time, bank"
        description:
          type: string
        requirements:
          type: ["string", "null"]
        salary:
          type: string
        postedAt:
          type: string
          format: date-time
        isActive:
          type: boolean
    NewsArticle:
      type: object
      required: [id, title, slug, excerpt, content, category, publishedAt, imageUrl]
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: string
        content:
          type: string
        category:
          type: string
        publishedAt:
          type: string
          format: date-time
        imageUrl:
          type: string
        author:
          type: ["string", "null"]
...